Explicit error message when dimensions requested exceeds the maximum permitted of a raster.
git-svn-id: http://svn.osgeo.org/postgis/trunk@9838 b70326c6-7e19-0410-871a-916f4a2858ee

+++ raster/rt_core/rt_api.h
@@ -765,7 +765,7 @@ rt_band rt_band_reclass(rt_band srcband, rt_pixtype pixtype,
  *
  * @return an rt_raster or 0 if out of memory
  */
-rt_raster rt_raster_new(uint16_t width, uint16_t height);
+rt_raster rt_raster_new(uint32_t width, uint32_t height);
 
 /**
  * Construct an rt_raster from a binary WKB representation
